website/requirements.txt
Flask==3.0.0
psycopg2-binary==2.9.9
Scrapy==2.11.0
scrapy-playwright==0.0.33